+++ src/components/FeaturePanel/ImagePane/Image/helpers.tsx
@@ -3,10 +3,10 @@ import React, { useEffect, useRef } from 'react';
 import Router from 'next/router';
 import { Size } from '../types';
 import { HEIGHT } from '../helpers';
-import { ImageDef, isTag } from '../../../../services/types';
 import { useFeatureContext } from '../../../utils/FeatureContext';
 import { getOsmappLink } from '../../../../services/helpers';
 import { removeFilePrefix } from '../../Climbing/utils/photo';
+import { ImageDef, isTag } from '../../../../services/types';
 
 export const initialSize: Size = { width: 100, height: HEIGHT }; // until image size is known, the paths are rendered using this (eg. ssr)
