+++ src/services/images/__tests__/getImageDefs.test.ts
@@ -1,4 +1,5 @@
 import { getImageDefs } from '../getImageDefs';
+
 import { ImageDef } from '../../types';
 
 const center = [14, 50];
